+++ test/_assertion_test.js
@@ -68,16 +68,4 @@ describe("END-TO-END: Assertion rendering", function() {
 		}, "'#element' doesn't have a property named 'XXX'. Did you misspell it?");
 	});
 
-	it("fails nicely when adding incompatible elements", function() {
-		assert.exception(function() {
-			element.width.plus(element.top).value();
-		}, "Size isn't compatible with Position");
-	});
-
-	it("fails nicely when diffing incompatible elements", function() {
-		assert.exception(function() {
-			element.width.diff(element.top);
-		}, "Can't compare width of '#element' to top edge of '#element': Size isn't compatible with Position");
-	});
-
 });